Introducing AbstractDiscriminator base class to avoid repeated start()/stop() implementation in each concrete Discriminator
